requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.1
Pillow==10.2.0
werkzeug==3.0.1
PyJWT==2.8.0 